Some web/backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-JWT-Extended==4.5.2
Flask-Migrate==4.0.4
Flask-CORS==4.0.0
python-dotenv==1.0.0
bcrypt==4.0.1
Pillow==10.0.0
python-dateutil==2.8.2 